trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
</head>
<body>
<script>
  run_tests_in_child_frame(
  '<meta http-equiv="Content-Security-Policy" content="trusted-types SomeName JustOneMoreName">',
  () => {
    // Whitelisted name test
    test(t => {
      let policy = window.TrustedTypes.createPolicy('SomeName', { createHTML: s => s } );
      assert_equals(policy.name, 'SomeName');
    }, "Whitelisted policy creation works.");

    // Another whitelisted name test
    test(t => {
      let policy = window.TrustedTypes.createPolicy('JustOneMoreName', { createHTML: s => s } );
      assert_equals(policy.name, 'JustOneMoreName');
    }, "Another whitelisted policy creation works.");

    // Non-whitelisted names test
    test(t => {
      assert_throws(new TypeError(), _ => {
       window.TrustedTypes.createPolicy('SomeOtherName', { createURL: s => s } );
      });
    }, "Non-whitelisted policy creation throws.");
  });
</script>
